code/ML_code/requirements.txt
google-cloud-contentwarehouse==0.4.0
google-cloud-documentai==2.13.0
google-cloud-storage==2.7.0
PyPDF2==3.0.1


